Filled in logic to use variant filtering
